Two tweaks/bug fixes for lecturing.
txt_mixin.py returns None if the list of indices was empty.  This was
causing a problem with returning out[0].  The reset lecture dialog of
tk_simple_dialog.py now handles a return on the go button and ses the
focus on the go button initially.  In gimp, typing Ctrl-Alt-1 and then
hitting return should run the reset script with the default settings.